pyproject.toml
[tool.poetry]
name = "snowflake-challenge-ml"
version = "0.1.0"
description = ""
authors = ["Peter Hansen <phansen@phdata.io>"]
readme = "README.md"
packages = [{include = "snowflake_ml"}]

[tool.poetry.dependencies]
python = ">=3.8.1, <3.11, !=3.9.7"
snowflake-snowpark-python = "1.7.0"
streamlit = "^1.27.1"
watchdog = "^3.0.0"
plotly = "^5.17.0"
snowflake-ml-python = "^1.0.8"
ydata-profiling = "^4.5.1"
python-dotenv = "^1.0.0"
shap = "^0.42.1"
xgboost = "1.7.3"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
